factored out visualization stuff to new module
